% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMBMkappa.R
\docType{package}
\name{PCMkappa}
\alias{PCMkappa}
\title{PCMBMKappa}
\description{
Add on for PCMBase stablishing the kappa class of evolutionary model to investigate the proportionality between matrices in a phylogenetic context.
}
