% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StationaryVariance.R
\name{StationaryVariance}
\alias{StationaryVariance}
\title{Stationary Variance of a OU process}
\usage{
StationaryVariance(H, Sigma_x)
}
\arguments{
\item{H}{Shur-factor of the adaptive rate matrix.}

\item{Sigma_x}{Upper triangular factor of the brownian rate matrix.}
}
\description{
Calculates the stationary varianve of a multivariate OU model.
}
